backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.1.4
numpy==1.26.4
implicit==0.7.2
joblib==1.3.2
pyarrow==14.0.2
python-dotenv==1.0.0
pydantic==2.5.3